refactor: group common build settings between modules in parent. move shared utils to common module
