+++ src/main/java/com/weirddev/testme/intellij/template/TemplateDescriptor.java
@@ -1,7 +1,7 @@
 package com.weirddev.testme.intellij.template;
 
 import com.weirddev.testme.intellij.icon.IconTokensReplacerImpl;
-import com.weirddev.testme.intellij.resolvers.groovy.LanguageUtils;
+import com.weirddev.testme.intellij.common.utils.LanguageUtils;
 import com.weirddev.testme.intellij.template.context.Language;
 
 /**
