+++ src/main/java/com/weirddev/testme/intellij/utils/PropertyUtils.java
@@ -3,7 +3,7 @@ package com.weirddev.testme.intellij.utils;
 import com.intellij.psi.PsiMethod;
 import com.intellij.psi.util.PropertyUtil;
 import com.weirddev.testme.intellij.resolvers.groovy.GroovyPropertyUtil;
-import com.weirddev.testme.intellij.resolvers.groovy.LanguageUtils;
+import com.weirddev.testme.intellij.common.utils.LanguageUtils;
 
 /**
  * Created with IntelliJ IDEA.
