+++ testme-intellij-groovy/src/main/java/com/weirddev/testme/intellij/resolvers/groovy/GroovyPsiTreeUtils.java
@@ -2,6 +2,7 @@ package com.weirddev.testme.intellij.resolvers.groovy;
 
 import com.intellij.psi.*;
 import com.intellij.psi.util.PsiTreeUtil;
+import com.weirddev.testme.intellij.common.utils.LanguageUtils;
 import com.weirddev.testme.intellij.resolvers.to.ResolvedMethodCall;
 import com.weirddev.testme.intellij.resolvers.to.ResolvedReference;
 import org.jetbrains.plugins.groovy.lang.psi.api.statements.arguments.GrArgumentList;
